+++ index.js
@@ -16,22 +16,31 @@ exports.register = function () {
         outbound = this.haraka_require('outbound');
     }
 
-    if (this.cfg.main.check_outbound) {
+    if (this.cfg.main.check_mail_from) {
         this.register_hook('mail', 'check_mail_from');
     }
 }
 
 exports.load_qmd_ini = function () {
-    const plugin = this
-    plugin.cfg = plugin.config.get('qmail-deliverable.ini', function () {
-        plugin.load_qmd_ini()
+    this.cfg = this.config.get('qmail-deliverable.ini', {
+        booleans: [
+            '+main.check_mail_from',
+            '*.check_mail_from',
+        ]
+    },
+    function () {
+        this.load_qmd_ini()
     })
+
+    if (this.cfg.main.check_outbound !== undefined) {
+        // backwards compat
+        this.cfg.main.check_mail_from = this.cfg.main.check_outbound
+    }
 }
 
 exports.check_mail_from = function (next, connection, params) {
-    const plugin = this;
 
-    if (!plugin.cfg.main.check_outbound) return next();
+    if (!this.cfg.main.check_mail_from) return next();
 
     // determine if MAIL FROM domain is local
     const txn = connection.transaction;
@@ -39,169 +48,178 @@ exports.check_mail_from = function (next, connection, params) {
     const email = params[0].address();
     if (!email) {
         // likely an IP with relaying permission
-        txn.results.add(plugin, { skip: 'mail_from.null', emit: true });
+        txn.results.add(this, { skip: 'mail_from.null', emit: true });
         return next();
     }
 
     const domain = params[0].host.toLowerCase();
 
-    plugin.get_qmd_response(connection, domain, email, (err, qmd_r) => {
+    this.get_qmd_response(connection, params[0], (err, qmd_r) => {
         if (err) {
-            txn.results.add(plugin, {err});
+            txn.results.add(this, {err});
             return next(DENYSOFT, err);
         }
 
         // the MAIL FROM sender is verified as a local address
         if (qmd_r[0] === OK) {
-            txn.results.add(plugin, { pass: `mail_from.${qmd_r[1]}` });
+            txn.results.add(this, { pass: `mail_from.${qmd_r[1]}` });
             txn.notes.local_sender=domain;
             return next();
         }
 
         if (qmd_r[0] === undefined) {
-            txn.results.add(plugin, { err: `mail_from.${qmd_r[1]}` });
+            txn.results.add(this, { err: `mail_from.${qmd_r[1]}` });
             return next();
         }
 
-        txn.results.add(plugin, { msg: `mail_from.${qmd_r[1]}` });
-        return next(CONT, `mail_from.${qmd_r[1]}`);
+        txn.results.add(this, { msg: `mail_from.${qmd_r[1]}` });
+        next(CONT, `mail_from.${qmd_r[1]}`);
     })
 }
 
-exports.get_next_hop = function (dom_cfg, queue_wanted) {
-
-    if (dom_cfg.next_hop) return dom_cfg.next_hop;
-
-    switch (queue_wanted) {
-        case 'lmtp':
-            return (`lmtp://${dom_cfg.host}`);
-        default:
-            return (`smtp://${dom_cfg.host}`);
-    }
-}
-
-exports.set_queue = function (connection, queue_wanted, domain) {
-    const plugin = this;
-
-    let dom_cfg = plugin.cfg[domain];
-    if (dom_cfg === undefined) dom_cfg = plugin.cfg.main;
-
-    if (!queue_wanted) queue_wanted = dom_cfg.queue;
-    if (!queue_wanted) return true;
-
-    const txn = connection.transaction;
-    if (!txn) return;
-    const next_hop = plugin.get_next_hop(dom_cfg, queue_wanted);
-
-    if (!txn.notes.get('queue.wants')) {
-        txn.results.add(plugin, {msg: `setting queue.wants=${queue_wanted}, ${next_hop}`})
-        txn.notes.set('queue.wants', queue_wanted);
-        txn.notes.set('queue.next_hop', next_hop);
-        return true;
-    }
-
-    // multiple recipients with same destination
-    if (txn.notes.get('queue.wants') === queue_wanted &&
-        txn.notes.get('queue.next_hop') === next_hop) {
-        return true;
-    }
-
-    // multiple recipients with different forward host, soft deny
-    return false;
+function do_relaying (plugin, txn, next) {
+    // any RCPT is acceptable for txns with relaying privileges
+    // this is called in several places where errors or non-local rcpt would
+    // otherwise not be allowed
+    txn.results.add(plugin, {pass: `relaying${txn.notes.local_sender ? ' local sender' : ''}`});
+    txn.notes.set('queue.wants', 'outbound');
+    next(OK);
 }
 
 exports.hook_rcpt = function (next, connection, params) {
-    const plugin = this;
     const txn = connection.transaction;
 
     const rcpt = params[0];
-    const domain = rcpt.host.toLowerCase();
-    let dom_cfg = plugin.cfg[domain];
-    if (dom_cfg === undefined) dom_cfg = plugin.cfg.main;
 
     // Qmail::Deliverable::Client does a rfc2822 "atext" test
     // but Haraka has already validated for us by this point
-    plugin.get_qmd_response(connection, domain, rcpt.address(), (err, qmd_r) => {
+    this.get_qmd_response(connection, rcpt, (err, qmd_res) => {
         if (err) {
-            txn.results.add(plugin, { err });
+            if (connection.relaying) return do_relaying(this, txn, next)
+            txn.results.add(this, { err });
             return next(DENYSOFT, "error validating email address");
         }
-
-        const [r_code, dst_type] = qmd_r;
-
-        // a client with relaying privileges is sending from a local domain.
-        // Any RCPT is acceptable.
-        if (connection.relaying && txn.notes.local_sender) {
-            txn.results.add(plugin, {pass: "relaying local_sender"});
-            plugin.set_queue(connection, 'outbound');
-            const q = txn.notes.get('queue.wants');
-            if (q) connection.loginfo(`queue: ${q}`);
-            return next(OK);
-        }
-
-        if (r_code === OK) {
-            txn.notes.local_recipient=domain;
-            txn.results.add(plugin, {pass: `rcpt.${dst_type}`, emit: true });
-            let queue = dom_cfg.queue;
-            if (dst_type === 'vpopmail dir' && dom_cfg.next_hop) {
-                if (/^lmtp/.test(dom_cfg.next_hop)) queue = 'lmtp';
-            }
-            if (plugin.set_queue(connection, queue, domain)) {
-                return next(OK);
-            }
-            return next(DENYSOFT, "Split transaction, retry soon");
-        }
-
-        if (r_code === undefined) {
-            txn.results.add(plugin, {err: `rcpt.${  dst_type}`});
-            return next();
-        }
-
-        // no need to DENY[SOFT] for invalid addresses. If no rcpt_to.* plugin
-        // returns OK, then the address is not accepted.
-        txn.results.add(plugin, {msg: `rcpt.${  dst_type}`});
-        return next(CONT, dst_type);
-    });
+        this.do_qmd_response(qmd_res, connection, rcpt, next)
+    })
 }
 
-exports.get_qmd_response = function (connection, domain, email, cb) {
+exports.do_qmd_response = function (qmd_res, connection, rcpt, next) {
+    const txn = connection.transaction;
+
+    const [r_code, dst_type] = qmd_res;
+
+    if (r_code === undefined) {
+        if (connection.relaying) return do_relaying(this, txn, next)
+        txn.results.add(this, {err: `rcpt.${dst_type}`});
+        return next();
+    }
+
+    if (r_code !== OK) {
+        if (connection.relaying) return do_relaying(this, txn, next)
+        // no need to DENY[SOFT] for invalid addresses. If no rcpt_to.* plugin
+        // returns OK, then the address is not accepted.
+        txn.results.add(this, {msg: `rcpt.${dst_type}`});
+        return next(CONT, dst_type);
+    }
+
+    const domain = rcpt.host.toLowerCase();
+    const dom_cfg = this.cfg[domain] || this.cfg.main;
+
+    txn.notes.local_recipient=domain;
+    txn.results.add(this, {pass: `rcpt.${dst_type}` });
+
+    let queue = this.get_queue(domain)
+    let next_hop = this.get_next_hop(domain, queue)
+
+    if (dst_type === 'vpopmail dir' && next_hop) {
+        if (/^lmtp/.test(next_hop)) queue = 'lmtp';
+        next_hop = this.get_next_hop(domain, queue)
+    }
+
+    if (this.is_split(txn, queue, next_hop)) {
+        if (dom_cfg?.split === 'defer') {
+            if (connection.relaying) return do_relaying(this, txn, next)
+            return next(DENYSOFT, "Split transaction, retry soon");
+        }
+        txn.results.add(this, {msg: `split queue.wants=outbound`, emit: true})
+        txn.notes.set('queue.wants', 'outbound');
+        delete txn.notes?.queue?.next_hop;
+    }
+    else {
+        if (!txn.notes.get('queue.wants')) {
+            txn.results.add(this, {msg: `queue.wants=${queue}, next_hop=${next_hop}`, emit: true})
+            txn.notes.set('queue.wants', queue);
+            txn.notes.set('queue.next_hop', next_hop);
+        }
+    }
+
+    next(OK);
+}
+
+exports.is_split = function (txn, queue, next_hop) {
+
+    if (txn.rcpt_to.length > 1) return true;
+
+    const qw = txn.notes.get('queue.wants')
+    if (qw && qw !== queue) return true
+
+    const qnh = txn.notes.get('queue.next_hop')
+    if (qnh && qnh !== next_hop) return true
+
+    return false  // identical destinations
+}
+
+exports.get_next_hop = function (domain, queue) {
+    const hop = this.cfg[domain]?.next_hop || this.cfg.main.next_hop
+    if (hop) return hop
+    return `${queue === 'lmtp' ? 'lmtp' : 'smtp'}://${this.get_host(domain)}`
+}
+
+exports.get_queue = function (domain) {
+    // lmtp, outbound, smtp_forward, qmail-queue
+    return this.cfg[domain]?.queue || this.cfg.main.queue;
+}
+
+exports.get_host = function (domain) {
+    return this.cfg[domain]?.host || this.cfg.main.host || '127.0.0.1'
+}
+
+exports.get_port = function (domain) {
+    return this.cfg[domain]?.port || this.cfg.main.port || 8998
+}
+
+exports.get_qmd_response = function (connection, addr, cb) {
     const plugin = this;
 
-    let cfg = plugin.cfg[domain];
-    if (cfg === undefined) cfg = plugin.cfg.main;
+    const domain = addr.host.toLowerCase()
+    const email  = addr.address()
 
     const options = {
         method: 'get',
-        host: cfg.host || '127.0.0.1',
-        port: cfg.port || 8998,
+        host: plugin.get_host(domain),
+        port: plugin.get_port(domain),
     };
 
-    // redundant
-    // connection.transaction.results.add(plugin, {
-    //     msg: "sock: " + options.host + ':' + options.port
-    // });
-
     connection.logdebug(plugin, `checking ${email}`);
-    options.path = `/qd1/deliverable?${  querystring.escape(email)}`;
+    options.path = `/qd1/deliverable?${querystring.escape(email)}`;
     // connection.logdebug(plugin, 'PATH: ' + options.path);
     http.get(options, function (res) {
-        connection.logprotocol(plugin, `STATUS: ${  res.statusCode}`);
-        connection.logprotocol(plugin, `HEADERS: ${  JSON.stringify(res.headers)}`);
+        connection.logprotocol(plugin, `STATUS: ${res.statusCode}`);
+        connection.logprotocol(plugin, `HEADERS: ${JSON.stringify(res.headers)}`);
         res.setEncoding('utf8');
         res.on('data', function (chunk) {
-            connection.logprotocol(plugin, `BODY: ${  chunk}`);
+            connection.logprotocol(plugin, `BODY: ${chunk}`);
             const hexnum = new Number(chunk).toString(16);
-            const arr = plugin.check_qmd_response(connection, hexnum);
+            const arr = plugin.decode_qmd_response(connection, hexnum);
             connection.logdebug(plugin, arr[1]);
             cb(undefined, arr);
-
         });
     }).on('error', cb)
 }
 
-exports.check_qmd_response = function (connection, hexnum) {
-    const plugin = this;
-    connection.logprotocol(plugin,`HEXRV: ${  hexnum}` );
+exports.decode_qmd_response = function (connection, hexnum) {
+
+    connection.logprotocol(this,`HEXRV: ${  hexnum}` );
 
     switch (hexnum) {
         case '11':
@@ -248,39 +266,41 @@ exports.check_qmd_response = function (connection, hexnum) {
 
 exports.hook_queue = function (next, connection) {
 
-    const wants = connection.transaction.notes.get('queue.wants');
-    if (wants !== 'lmtp') return next();
-
-    this.logdebug('QMD routing to outbound');
-    outbound.send_email(connection.transaction, next);
+    const qw = connection.transaction.notes.get('queue.wants')
+    switch (qw) {
+        case 'lmtp':
+        case 'outbound':
+            this.loginfo(`routing to outbound: queue.wants=${qw}`);
+            outbound.send_email(connection.transaction, next);
+            break;
+        default:
+            next(); // do nothing
+    }
 }
 
 exports.hook_get_mx = function (next, hmail, domain) {
-    const plugin = this;
 
     if (hmail.todo.notes.get('queue.wants') !== 'lmtp') return next();
-
-    let cfg = plugin.cfg[domain.toLowerCase()];
-    if (cfg === undefined) cfg = plugin.cfg.main;
 
     const mx = {
         using_lmtp: true,
         priority: 0,
         port: 24,
-        exchange: cfg.host,
-    };
+        exchange: this.get_host(domain.toLowerCase()),
+    }
 
-    if (cfg.next_hop) {
-        const dest = url.parse(cfg.next_hop);
-        if (dest.hostname) mx.exchange = dest.hostname;
-        if (dest.port) mx.port = dest.port;
+    const nh = hmail.todo.notes.get('queue.next_hop')
+    if (nh) {
+        const dest = new url.URL(nh)
+        if (dest.hostname) mx.exchange = dest.hostname
+        if (dest.port) mx.port = dest.port
         if (dest.auth) {
-            mx.auth_type = 'plain';
-            mx.auth_user = dest.auth.split(':')[0];
-            mx.auth_pass = dest.auth.split(':')[1];
+            mx.auth_type = 'plain'
+            mx.auth_user = dest.auth.split(':')[0]
+            mx.auth_pass = dest.auth.split(':')[1]
         }
     }
 
-    plugin.logdebug(mx);
-    return next(OK, mx);
+    this.logdebug(mx)
+    next(OK, mx)
 }